factor multisig functions in its own helper
