+++ smart-contracts/scripts/upgrade/propose.js
@@ -1,62 +1,11 @@
-const { ethers, network } = require('hardhat')
+const { ethers } = require('hardhat')
 
-const multisigABI = require('../../test/helpers/ABIs/multisig.json')
-const proxyABI = require('../../test/helpers/ABIs/proxy.json')
-
-// TODO: make possible to switch to DAO
-const unlockMultisigAddress = '0xa39b44c4AFfbb56b76a1BF1d19Eb93a5DfC2EBA9'
-
-const confirmTx = async ({ signers, multisig, transactionId }) => {
-  const txs = await Promise.all(
-    signers.slice(1, 4).map(async (signerAddress) => {
-      await network.provider.request({
-        method: 'hardhat_impersonateAccount',
-        params: [signerAddress],
-      })
-      const balance = ethers.utils.hexStripZeros(
-        ethers.utils.parseEther('1000')
-      )
-      await network.provider.send('hardhat_setBalance', [
-        signerAddress,
-        balance,
-      ])
-
-      const signer = await ethers.getSigner(signerAddress)
-
-      const m = multisig.connect(signer)
-      const tx = await m.confirmTransaction(transactionId, {
-        gasLimit: 1200000,
-      })
-      return await tx.wait()
-    })
-  )
-  const [lastTx] = txs.slice(-1)
-  const { events, transactionHash } = lastTx
-  const failure = events.find((v) => v.event === 'ExecutionFailure')
-  const success = events.find((v) => v.event === 'Execution')
-
-  if (failure) {
-    // eslint-disable-next-line no-console
-    console.log(
-      `ERROR: Proposal ${transactionId} failed to execute (txid: ${transactionHash})`
-    )
-  } else if (success) {
-    // eslint-disable-next-line no-console
-    console.log(
-      `Proposal ${transactionId} executed successfully (txid: ${transactionHash})`
-    )
-  }
-}
-
-const encodeUpgradeTxData = async ({ proxyAddress, implementation }) => {
-  // build upgrade tx
-  const proxy = await ethers.getContractAt(proxyABI, proxyAddress)
-  const data = proxy.interface.encodeFunctionData('upgrade', [
-    proxyAddress,
-    implementation,
-  ])
-  return data
-}
+const {
+  getUnlockMultisig,
+  encodeUpgradeTxData,
+  confirmMultisigTx,
+} = require('../../helpers/multisig')
+const { impersonate } = require('../../test/helpers/mainnet')
 
 // used to update contract implementation address in proxy admin using multisig
 async function main({ proxyAddress, proxyAdminAddress, implementation }) {
@@ -66,11 +15,8 @@ async function main({ proxyAddress, proxyAdminAddress, implementation }) {
   // eslint-disable-next-line no-console
   if (isDev) console.log('Dev mode ON')
 
-  // get multisig
-  const multisig = await ethers.getContractAt(
-    multisigABI,
-    unlockMultisigAddress
-  )
+  // TODO: make possible to switch to DAO
+  const multisig = await getUnlockMultisig()
   const owners = await multisig.getOwners()
 
   // get proper credentials
@@ -78,13 +24,7 @@ async function main({ proxyAddress, proxyAdminAddress, implementation }) {
   if (isDev) {
     // impersonate multisig owner
     issuer = await ethers.getSigner(owners[0])
-    await network.provider.request({
-      method: 'hardhat_impersonateAccount',
-      params: [issuer.address],
-    })
-    // give some ETH
-    const balance = ethers.utils.hexStripZeros(ethers.utils.parseEther('1000'))
-    await network.provider.send('hardhat_setBalance', [issuer.address, balance])
+    await impersonate(owners[0])
   } else {
     ;[issuer] = await ethers.getSigners()
   }
@@ -117,7 +57,7 @@ async function main({ proxyAddress, proxyAdminAddress, implementation }) {
 
   if (isDev) {
     // reach concensus
-    await confirmTx({ signers: owners, multisig, transactionId })
+    await confirmMultisigTx({ transactionId })
   }
   return transactionId
 }
